OMAP3: PM: remove WARN() on PRCM IRQ status bits
Status bits in PRM_IRQSTATUS_MPU will still be set by hardware even if
the corresponding IRQ enable bits in PRM_IRQENABLE_MPU are not.  This
led to false positives in prcm_interrupt_handler.

For example WK_ST might be set and cleared correctly by
prcm_interrupt_handler, but that same interrupt service would catch
stray set bits (ie, VP1_TRANXDONE_ST) for which no IRQ was enabled.
This either had to be handled manually (catch and clear all set bits for
which no work was to be done) or resulted in a WARN() for an unhandled
interrupt.

Patch removes the manually caught and cleared bits for which no work is
performed as well as the WARN().  The downside is that legitimate
interrupts that are not handled are silent, but this should not happen
and the false positives were very confusing and noisy.

Change-Id: I9e17b2f6b06e83e4a7da48089d244cea400ed095
Signed-off-by: Mike Turquette <mturquette@ti.com>
Acked-by: Girish S G <girishsg@ti.com>