USB: MUSB: Enable DMA mode1 RX for gadget android and gadget file storage
This patch enables the DMA mode1 RX support for the
android and file storage gadget.

This patch does a runtime check of the loaded gadget
and enables mode1 DMA RX if the gadget is g_android
or g_file_storage

This will result in a thpt performance gain of around 30%
for USB mass-storage use cases.

Change-Id: I23fde195695344a3ed96bd370d4a81213e0dabc2
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>
Acked-by: Vikram Pandita <vikram.pandita@ti.com>