OMAP:defconfig change to make android gadget built in to kernel
Android gadget is written to be built in to kernel. The zoom2 and
zoom3 defconfigs were building this as module, which caused instabilities
in the system. The following were observed:
1. Sometimes android gadget doesn't work until is mounted two times
in a row on device.
2. If the gadget module is removed and reinserted, a kernel crash
occurs.

This patch fixes the issue by changing the defconfig to build the
gadget driver in to the kernel.

Signed-off-by: Sudheesh J <sudheesh.joseph@ti.com>