UART : Errata fix for MDR1 access can Freeze UART Module
This patch fixes the Errata 1.6 on 3630 and 1.12 on 3430. The MDR1
access is done at every call of omap_uart_restore_context, from the
omap_sram_idle. This was causing the system to crash after few cyc-
les of the MDR1 access.

Description : From Errata Doc.
Because of a glitchy structure inside the UART module, accessing the
MDR1 register may create a dummy underrun condition and freeze the
UART in IrDa transmission. In UART mode, this may corrupt the trans-
mitted data.

WORKAROUND
To ensure this problem does not occur, the following software initi-
alization sequence must be used each time MDR1 must be changed:
- If needed, setup the UART by writing the required registers, exc-
ept MDR1
- Set appropriately the MDR1.MODE_SELECT bit field
- Wait for 5 L4 clock cycles + 5 UART functional clock cycles
- Clear TX and RX FIFO in FCR register to reset its counter logic

Signed-off-by: Deepak K <deepak.k@ti.com>
Acked-by: Govindraj Raja <govindraj.raja@ti.com>