TILER: Added reference counting and cleanup. Fixed query bugs.
The reference count of blocks increases on alloc, and by registered
buffers that contain that block.

When all tiler drivers of a process are closed:
  - blocks allocated by it are marked "freed"
  - buffers registered by it are unregistered
All blocks that are no longer marked "allocated" and are not
referenced by registered buffers are actually freed.

Blocks allocated by the kernel API are tracked separately, and
are not release when tiler drivers are closed.  They are only
freed using tiler_free, or when the kernel driver closes.

Fixed a bug with querying buffers by address, which did not
account for the new TCM support for 1D buffers and block management.

Added support for clearing PAT for freed blocks.

Added extended tiler kernel API-s for specifying group ids, pids,
alignment and offset.

tiler_allocx
tiler_reserve
tiler_reservex
tiler_map
tiler_mapx

Signed-off-by: Lajos Molnar <molnar@ti.com>