USB: MUSB: Enable DMA mode1 RX for USB-mass-storage
This patch enables the DMA mode1 RX support for the
android and file storage gadget.

This patch does a runtime check of the loaded gadget
and enables mode1 DMA RX if the gadget is g_android
or g_file_storage

This will result in a thpt performance gain of around 30%
for USB mass-storage use cases.

Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>