OMAP: MUSB: Prevent CORE OFF when cable attached for OMAP3630
Prevent the CORE domain from going to OFF mode when USB cable is attached
MUSB does not have h/w SaveAndRestore(SAR) and hence cannot work with
core off.

This patch restricts CPUIdle from going to C7 state.
On usb cable attach, a constrant is put on CORE not to go to off and
on cable detach the constraint is removed.

Without this patch, if OFF mode is enabled, even with usb cable attached and
device enumerated, there can be device disconnection.

REVIST for OMAP4 after PM implementation

Signed-off-by: Vikram Pandita <vikram.pandita@ti.com>
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>