ZOOM3: Request backlight gpio after twl init.
GPIO199 is part of twl, when requesting for this gpio
before twl init, the gpio request would fail, because
the chip was not registered in the gpio_desc table.
Eventuallly, it  would be autorequested the first time
it was used, producing an ugly warning.

Moving the request will ensure that the gpio is
requeseted after twl has registerd in gpio_desc.

Signed-off-by: Axel Haslam <axelhaslam@ti.com>