OMAP: I2C: hacking i2c rev detection for omap3/4
OMAP3 initialization is showing OMAP_I2C_REV value of 0x40,
which is declare to be the value for OMAP4 platforms. This
is resulting in a kernel instability.

By keeping OMAP4 check greater that 0x40, the problem is
resolved, but not properly fixed. Let's keep this as a
temporary __HACK__

Signed-off-by: Leed Aguilar <leed.aguilar@ti.com>