USB: MUSB: Fallback to PIO mode if DMA address is not aligned
On the new MUSB controller,(RTL rev 1.8 on OMAP3630 and
RTL rev 2.0 on OMAP4430) DMA does not work properly if
DMA address is not alligned on 4 byte.

Make sure the DMA address is 4 byte aligned, if not we
use the interrupt (PIO) mode instead of using DMA mode
for OMAP3630 and beyond.

Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>