OMAP3: I2C: fix check i2c revision for 3630 arch
I2C revision on OMAP4 and OMAP3630 are returning the
same revision number: 0x40. The current logic in the
driver initialization and idle path to decide which
action to take based in the OMAP architecture was
causing I2C to fail for one of the board.

Keeping this check helps to prevent an i2c failure.

Signed-off-by: Leed Aguilar <leed.aguilar@ti.com>