USB: MUSB: Update DMA buffer address between DMA/PIO transfers
This patch fixes the issues inroduced by following patch:
"Do not enable TX and RX DMA at the same time"

Because the above patch dynamically use DMA or PIO for TX/RX
transfers based on the direction, DMA buffer address may not
be updated if a PIO is used between DMA packet transfers.

This patch updates the DMA buffers accordingly.

Signed-off-by: Wenbiao Wang <wwang@ti.com>
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>