omap3isp: Avoid errata i421 for AE and AF paxel size
This avoids a corner case, in which an ISP errata is hit.

The errata consists on wrong H3A output when an AF window ends
2 pixels before AE window ends. In that case, the unsaturated
accumulator value is zeroed for that entry, leading to wrong data.

So, do a validation on the paxel size relationship, to see
if that will be hit or not somewhere during the stats recollection.

For more information, refer to omap3630 errata document v0.4, entry
"1.40. ISP H3A wrong data output".

Change-Id: Ife5b5741f1d7269a7ab95b3f784164d3be8dae90
Signed-off-by: Sergio Aguirre <saaguirre@ti.com>
Acked-by: Christina Warren <cawarren@ti.com>
Acked-by: Akwasi Boateng <akwasi.boateng@ti.com>
Acked-by: Leed Aguilar <leed.aguilar@ti.com>