omap3isp: Avoid errata i421 for AE and AF paxel size
This avoids a corner case, in which an ISP errata is hit.

The errata consists on wrong H3A output when an AF window ends
2 pixels before AE window ends. In that case, the unsaturated
accumulator value is zeroed for that entry, leading to wrong data.

So, do a validation on the paxel size relationship, to see
if that will be hit or not somewhere during the stats recollection.

For more information, refer to omap3630 errata document v0.4, entry
"1.40. ISP H3A wrong data output".

NOTE: This is a draft meant for testing, and NOT a final version.

Signed-off-by: Sergio Aguirre <saaguirre@ti.com>