SYSLINK: IPC: Make Syslink Loadable Modules
This patch has the kernel changes to remove the circular dependency
and make syslink into loadable modules.  To achieve this, one new
module is added, which was seperated from the syslink_ipc module.
The new module is syslink_platform and contains  multiproc and platformmem
code.  This new module does not depend on any other module to load.
There are corresponding kernel build file changes and userspace-syslink
changes needed in conjunction with this change.

Change-Id: Ic8724915100cde02c0c93f82c1b49125f207d67c
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>