UART : fix crash for BT Transaction with Core OFF enabled
This fixes the crashes happening while the system exits out of Core
OFF mode. While the audio is being played, the system enters sleep
mode when the Pause button is pressed on BT headset. When the re-
sume of pressed the system exit core off mode. It can happen due to
the DMA scenario the ICLK or FCLK are disabled while transaction is
going on for a very small period ( can cause a crash ). To avoid
this scenarios the clocks are enable while a Serial IRQ occurs and
while the DMA TX call back is done.

The wait lock is being held in DMA tx callback since at the end of
transaction, there is small amount of handsacking is going on with
the BT headset, hence to settle down, this wait lock is being held.

Signed-off-by: Deepak K <deepak.k@ti.com>
Acked-by: Govindraj Raja <govindraj.raja@ti.com>
Acked-by: Mike Turquette <mturquette@ti.com>