LEDS: leds-pwm: Add init, notfiy and exit callbacks
This patch adds init, notify and exit callbacks to the leds-pwm driver.

One usecase for these callbacks is that on certain platforms it is neccessary
to configure a gpio pin as pwm pin. On these platforms it is usefull to have
the added callbacks, so that the gpio pin can be configured at the same time
as the pwm device.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>