EHCI: force the domain to Idle/Standby while disabling the f-clocks
if the USB F-clocks are disabled while the domain is active, we
will not be able to activate the domain again. Workaround this
by forcing the domain to Idle/Standby while disabling the f-clocks.
The root cause of this issue is not yet known, so this is a
temporary workaround to prevent a kernel panic.
We force-idle, force-standby the controller before disabling the
f-clocks and put it back to no-idle, no-standby after enabling
the f-clocks.

Signed-off-by: Axel Haslam <axelhaslam@ti.com>
Acked-by: Vikram Pandita <vikram.pandita@ti.com>