TWL4030: Battery: Notify user of USB cable plugged.
USB plug-in does not generate a interupt on the
battery Driver. The USB interrupt is handeled
by the USB driver which calls the usb_en function
in this battery driver.

Since in usb_en we are not calling power_supply_changed
to notify user space of the plug in event, We wait
until the next update cycle to detect the change in
power_suply and send the event to user space.

Change-Id: I0077ad17b51af1112dcb618c831f16d6ec2ac3f2
Signed-off-by: Madhusudhan Chikkature <madhu.cr@ti.com>
Signed-off-by: Axel Haslam <axelhaslam@ti.com>
Acked-by: Leed Aguilar <leed.aguilar@ti.com>