TWL4030: Battery: Add sysfs entry for USB battery
Android battery service requires a sysfs folder on
/sys/class/power_supply that has a battery with the
"type" property POWER_SUPPLY_TYPE_USB.
The battery service uses the "online" property of
this battery to determine if the USB cable is
plugged in or not.

In turn, The USB activity Service is notified
by the battery service if the cable is plugged.

Without this sysfs entry, USB activity would
not get properly notified of the status of the
USB cable.

Change-Id: I9fbf863fc3004f60989fc9a5ba2347022ad7c4c3
Signed-off-by: Madhusudhan Chikkature <madhu.cr@ti.com>
Signed-off-by: Axel Haslam <axelhaslam@ti.com>
Acked-by: Leed Aguilar <leed.aguilar@ti.com>