TWL: BATTERY: update status before uevent
This is to update the status before calling
power supply changed. Without this, the
status is only updated on the work function,
and the result is that when we notify
user of power supply changed, on the irq
handler, the work has not yet updated the
status, and any action user takes based on
the sysfs entries at that time may not be correct.
Example: update battery status icon.

Change-Id: Ia72c1cb2e4144e3da2407b315db6f5b334118445
Signed-off-by: Axel Haslam <axelhaslam@ti.com>
Acked-by: Leed Aguilar <leed.aguilar@ti.com>