TWL: BATTERY: Modify capacity/voltage  values
When a cable is plugged in the battery symulator,
the reported voltage is 3.5v, with the current capacity
values, the result is 5% and would show up as red
on android battery indicator.
The voltage range of a operating battery is 4.2 to 3.2,
so this the percentege values to that range.
with the new values, 3.5 will show 50% charge at 3.5v,
which is more acurate

Change-Id: I01a853425b86ae4209b9e3b60f7683ad0bcceeb2
Signed-off-by: Axel Haslam <axelhaslam@ti.com>