OMAP3: v4l2: fix queued buffer state for proper queue cleanup
1. Queued buffers should in VIDEOBUF_QUEUED state so that it can
be properly deleted from queue when streamoff is called.
Currently the buffer state is set to VIDEOBUF_PREPARED when queued
and the buffers are not getting cleaned up when streamoff is called.
Only the buffers that are queued with state VIDEOBUF_QUEUED are
removed from the queue.
2. This is resulting in one of the buffers from being removed from the
queue and is getting dropped from being used by the Q-DQ cycle between
camera/v4l2 driver.
3. Setting the queued buffer state to VIDEOBUF_QUEUED fixes this issue.

Signed-off-by: Srinivas Pulukuru <srinivas.pulukuru@ti.com>