USB: MUSB: fix enumeration on windows host
TI product id 0xffff corrosponds to a bluetooth device.
This causes the windows host to send bluetooth SETUP message that caused
the enumeration to break.

Change the Product id to a non-allocated value.
Refer this link for allocated PIDs for TI VID=0451
	http://www.linux-usb.org/usb.ids

Tested with Windows laptop and Linux Fedora machine

Signed-off-by: Vikram Pandita <vikram.pandita@ti.com>
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>