omap: iommu- fix issues when building as modules
The function iommu_get_plat_data_size is used outside of the iommu
and needs to be exported when building as modules.  Also, the omap_iommu_class
is created in the module init in iommu.c but isn't destroyed in the module
exit which prevents the ability to repeatedly load and unload the module.
These issues are fixed in this patch.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>