OMAP3630: DSS: don't disable iclk twice in suspend
On OMAP3630 a data abort was happening due to patch 63b172a, which was
reverted by 743d382.  Issue was that during suspend path DPI code was
writing to DSS registers after iclks were already cut.

Best option is to combine the two patches to make OMAP3 and OMAP4 folks
happy.

Change-Id: I0f2056032140865616a7d2c90ab9476a95714f7a
Signed-off-by: Mike Turquette <mturquette@ti.com>
Acked-by: Leed Aguilar <leed.aguilar@ti.com>