OMAP3:PM: USB OTG Force idle first time kernel bootup
first time with kernel boot up without USB cable connected
need to force idle OTG, this is inline with the previous
commit 43ba01396e6fff9ce2431ab374ee017de38733dd
do:
 - smartsdby
 - forceidle

History: On 3630 all the USB autoidle issue was supposed
to be fixed. But unfortunately there were instances seen
that still OTG doesn't idle-ack as requested on some of
the SOM's. To have a robust solution;
forceidle/forcestdby when cable is detached and smartidle
when cable attached, is used to get suspend/resume and idle
path CORE RET/OFF working.

REVISIT: Need to Validate for ZOOM2 3430. And port relevant
patches from 2.6.29 to 2.6.32 as needed.

Signed-off-by: Girish S G <girishsg@ti.com>