glamo-mci: add a "nonremovable" field to glamo_mmc_platform_data
On the FR, the µSD can't be removed without removing the battery first.
So, we can safely assume that the µSD is not removable.

This patch make the MMC subsystem know that the card isn't removable if the new
"nonremovable" field in the platform data is true.

This affects, amongst other things, the suspend/resume process, the default
behaviour being to remove the card (after syncind any mounted FS) on suspend,
and re-detect it on resume, to prevent errors when removing the card while
the device is suspended.