OMAP: V4L2: Add PM constraints to video driver
This patch adds VDD1 and VDD2 constraints for video
driver. These constraints are set while streamon
is running, , in order to request the minimum
resources to ensure video playback performs properly

1. When stream is ON, VDD1 is set to 500Mhz (VDD1_OPP3)
for OMAP34xx and 600MHz (VDD1_OPP2) for OMAP3630
in order to enhance video playback performance.
Bus frequency is changed to 166 MHz for OMAP34xx
and 200 MHz for OMAP3630.

2. When stream is OFF, VDD1 is set to VDD1_OPP1 that
operates at 125Mhz for 3430 and 300MHz for 3630;
bus frequency constraint is also released.

Change-Id: Ie165c54a199d994f111b36de9617fee548932c96
Signed-off-by: Axel Castaneda Gonzalez <x0055901@ti.com>
Acked-by: Girish S G <girishsg@ti.com>