OMAP3:PM: Bug in cpuidle while checking proper CORE state
CORE OFF should be allowed only if all the peripheral domain are OFF.

However, the check for core_state was wrong. This caused CORE OFF to be
set even though other peripherals are active. Causing context lose for
some domains as CORE OFF transition was causing reset to that domain.

This is fixed by putting the right check.

Change-Id: I7762f504c5f9b645bfbdcef35d1db4db8cb7e061
Signed-off-by: Girish S G <girishsg@ti.com>
Acked-by: Mike Turquette <mturquette@ti.com>
Acked-by: Rajeev Kulkarni <rajeevk@ti.com>