SYSLINK:IPU-PM: change api for regulators
New regulators API available via ipu_pm resources FW
previosly a hack was being used to get and rel regulators
now the properly API functions are in use for the same
purpose.

Cam2Pwr is enable by default on boot time with a particular
voltage; when ipu_pm_get_regulator() is called the voltage is
changed to the value provided by requester. In order to release
the default voltage is set back to Cam2Pwr regulator.

Change-Id: I05b07d0c014b7d0ac0bfceeb5333c03a752dd57f
Signed-off-by: Miguel Vadillo <vadillo@ti.com>
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Paul Hunt <hunt@ti.com>