SYSLINK: ipc- adapt gatehwspinlock module to use hwspinlock driver
The gatemp, gatehwspinlock & gatepeterson modules have been enhanced
to support reserved resources for each gate_type.

This change is done as part of the gatehwspinlock adaptation to the
open-source hwspinlock driver code. gatemp module was managing all
the spinlocks previously, and with this patch, gatemp module will
only control and manage a contiguous block of spinlocks from 4 to 31.
The h/w spinlocks 0 to 3 will be reserved for usage by I2C driver.

All the spinlocks except the reserved spinlocks are acquired upfront
in the gatehwspinlock_setup function, instead of requesting it during
create time. This ensures the GateMP module has control over these so
that BIOS-side functionality can remain the same as before.

Change-Id: If520c58d18281ae2ffc274dda5aac72cf991aae2
Signed-off-by: Suman Anna <s-anna@ti.com>