OMAP-SERIAL: remove the "-1" on id parameter.
the id parameter is form 0 to 2 for each of
the omap uarts. if we substract 1 from
say, ttyO0, that would mean a id of -1,
which results in no traces at all on
3630sdp.

Also, this errata is only for omap3, this
adds checks so that it is not used in case
we are running on omap4.

Change-Id:	Iaee2e3d437abdb7416226aebbaa151ea1ce2bf41
Signed-off-by: Axel Haslam <axelhaslam@ti.com>