SYSLINK:IPU-PM: workqueue implementation for ipu_pm
Introducing the use of a workqueue to manage the request
and release of the resources using the ipu_pm module.
The implementation is using a FIFO to store the messages
(proc_id and payload) in the CB and enqueueing the work,
currently only one, then the work function is calling the
particular API to req/rel a resource.

Due to this implementation also some global variables were
changed to local.

FIFO depth is 16 but configurable.
Request and release API are now separated.

This implementation is based in the patch provided by
Fernando Guzman Lugo.

Signed-off-by: Miguel Vadillo <vadillo@ti.com>
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Paul Hunt <hunt@ti.com>
Signed-off-by: Fernando Guzman Lugo <x0095840@ti.com>

[Amend] Shivananda Hebbar <x0hebbar@ti.com>
Changes done for Kernel 32:

Use kfifo_alloc prototype based on k32, Use kfifo_get and kifo_put,
Remove spinlock protection as its taken care by the kfifo_get/put implimentations
internally.

Change-Id: If602f3c2546e376359f3897153fc2f31415b0332
Signed-off-by: Shivananda Hebbar <x0hebbar@ti.com>