SYSLINK: ipc - memory map changes to support new ducati heap
The Ducati memory map is adjusted to make more room for a heap buffer
by optimizing the memory used for different sections, and by reclaiming
the unused memory in the I/D bus to a heap. This patch makes the
corresponding changes in the Platform and DucatiEnabler code. This requires
a corresponding user-space change as well.

Change-Id: Ia6716971cb026f2474c7eeb94caede1668b889be
Signed-off-by: Suman Anna <s-anna@ti.com>