OMAP3:HQ: resizer: Enable to setup "ISP resizer" format together with different crop size.
The patch enabled more that 4 times difference between in/out "ISP resizer" images,
without automatic modification of output size. The condition is enabled when the
user set crop and input together. If the crop is small/large enough to produce
required output image without modification, the condition is fulfill.
That feature required by, "HQ-ICapture" and enable possibility to produce Preview
and Snapshot image with equal field of visibility.

Change-Id:Idc92713ce4e330ca0512337e7fd27cdfe5019f8e
Signed-off-by: Atanas Filipov <afilipov@mm-sol.com>
Signed-off-by: Akwasi Boateng <akwasi.boateng@ti.com>
Acked-by: Christina Warren <cawarren@ti.com>
Acked-by: Muthu Subramaniam <muthu@ti.com>