DSPBRIDGE: Fix wrong prints of DBC_ENSURE()
The patch "DSPBRIDGE: Check pointer instead of using MEM_IS_VALID_HANDLE
macro" uncovered an issue when calling the assertion macros after the
pointer was freed, but not set to NULL.

This issue was not showing before the above patch because the macro
MEM_IS_VALID_HANDLE checked for the pointer and signature, which was
not good because it dereferenced a pointer that was already freed.

This patch removes DBC_ENSURE() whenever it checks for a previously
freed pointer, or fix the code to avoid wrong assertions printing.

Signed-off-by: Ivan Gomez Castellanos <ivan.gomez@ti.com>
Signed-off-by: Rene Sapiens <rene.sapiens@ti.com>