omap:iommu-notifier callback mechanism for mmu events
Implement notifier callback mechanism for IOMMU events
such as MMU fault and file close.
The Error handler that hooks to this registration would notify
the remote processor whenever a process closed the iommu. This is
to ensure that the Processes on Remote process stop accessing the
buffers mapped by the terminating Process

Change-Id: I2854609c2cf81eae60e0ca9ae7daf9b09e7c2d87
Signed-off-by: Hari Kanigeri <h-kanigeri2@ti.com>
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>