bh1780: Reduce suspend and resume time for ALS sensor
Remove peripherial writes and reads from the suspend and resume routines
as reading the state of the device prior to writing it in the suspend routine
takes time.  Since the current power state is already stored save it off into another
variable and turn off the device.

Then when resuming check the saved off variable and either turn on the device or
leave it in the state that it was in before suspending.

This in turn will reduce the overall suspend/resume time.

Change-Id: I45420e05f184db7318edc6a9ceded2004b429633
Signed-off-by: Dan Murphy <dmurphy@ti.com>