ASoC: OMAP: Fix dma parameters overwritten in multistream scenarios
OMAP ABE DAI driver used a single array to save dma parameters (playback
and capture). When multiple streams are opened on different ports but
on the same direction (playback/capture), dma parameters are incorrectly
overwritten. That's prevent using an array with multiple entries, one for
each port supported in the DAI driver.

Change-Id: I355e3c9e7ad97c28cb7986bc315b79bd3f484991
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>