omap:iommu-disable the MMU interrupts in fault ISR
Instead of disabling the MMU device, just disable the MMU interrupts
when the MMU fault occurs. This change is required because the user
applications might be releasing the MMU resources after the MMU fault
recovery and this requires MMU to be in enabled state. MMU will be
disabled after all the open handles to IOMMU are closed.

Change-Id: Iec83e9f4ab37050a54065d625ef0d4552c7cd24a
Signed-off-by: Hari Kanigeri <h-kanigeri2@ti.com>
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>