DSPBRIDGE: Include missing info in MMU Fault debugging trace
Some execution info is missing at the beginning of the
trace buffer printout. This fix is to include this information
according to the buffer contents description and
also couple of cosmetic changes to make the traces more
readable.

This fix also uses DYNEXTBASE address to filter symbols that
match low addresses in the hash table and are not useful to
print. Any address below DYNEXTBASE is not valid for the DSP.

Change-Id: Ica87bba6e69199837b99c8c22bc9a830fcb6ee94
Signed-off-by: Ernesto Ramos <ernesto@ti.com>
Signed-off-by: Rene Sapiens <rene.sapiens@ti.com>