OMAP: OMAP-SERIAL: return 0 if checking for activity QUART
on checking for activity before suspend, the quart port
is hardcoded since it is not part of the OMAP uart list.
If the zoom is booted without a debugboard, the pdev for
quart is null and port 3 does not exist on the ui[] array
of omap-serial. this causes a null pointer dereference and
a board reset when trying to go into suspend in IDLE.

returning 0 if the port is not found on the list, since
this uart does not exist anyway.

Other possible solutions:
- check for debugboad precence before calling omap_uart_active.
- populating pdev even without a debugboard present.
- set a global flag stating if we have debugboard, and cheking it.

Signed-off-by: Axel Haslam <axelhaslam@ti.com>