OMAP3:Zoom2/3: Enable support for broken emmc.
The high capacity eMMC connected to MMC2 on the Zoom2/3 platform
reports a wrong EXT_CSD_REV. Due to this the MMC core does not
setup access mode correctly. This results in DATA CRC errors upon
a write access making the card unuseble.

This patch provides a __HACK__ to workaround this issue and makes
the card usable on the Zoom2/3 platforms having this card level issue.

The workaround is made specific to Zoom2/3 MMC2 under the configurable
option BROKEN_EMMC_ZOOM2_3. The MMC2 support is kept disabled by default.
Enabling MMC2 would select this config option for Zoom2/3.

Change-Id: I6c4529650c83092204e320f4b30d79ef9ba1ed9b
Signed-off-by: Madhusudhan Chikkature <madhu.cr@ti.com>