OAMP3:UART : Correct the UART 1 and 2 Pad Configs for wake-up
The Pad config values for UART 1 and UART 2 were correct to the
right location. If a wake happens, in the resume path would clear
the status bits. It can happen due to the DMA scenario the ICLK
or FCLK are disabled while transaction is going on for a very
small period ( can cause a crash ). To avoid this scenarios the
clocks are enable while a Serial IRQ occurs and while the DMA TX
call back is done.

The wait lock is being held in DMA tx callback since at the end of
transaction, there is small amount of handsacking is going on with
the BT headset, hence to settle down, this wait lock is being held.

Enabled Only UART 2 as default DMA mode for BT scenario
Added the save and restore of mcr register for the idle path fix
Corrected the DMA trigger values for errata fixes

Change-Id: I94a7819ba824fd00e3ad7c4c21fd88fa8acc2b74
Signed-off-by: Deepak K <deepak.k@ti.com>