OMAP3 Security Middleware Component 2.5.1 Driver
This patch is to integrate SMC [Security Middleware Component]
2.5.1 normal world driver to 2.6.32 for OMAP3.

This driver is not enabled by default . choose from menuconfig>>
device drivers > Security Middleware component to enable.

SMC provides a highly portable environment for running
security-sensitive code in a service-oriented fashion;
client applications executing in the Normal World make use of
functionality provided by security services executing in the
Secure World.

This version of the SMC is designed to run on OMAP3 processors,
which is based on a Cortex-A8 processor (v7 architecture).
This architecture provides isolation between Normal World and
Secure World memory and devices

Change-Id: I8a9c612a25f5bd35a0cd6467f16163cb36c1f7df
Signed-off-by: Praneeth Bajjuri <praneeth@ti.com>
Signed-off-by: Lionel Pacoud <lionel.pacoud@trusted-logic.fr>
Signed-off-by: Julien Deplancke <julien.deplancke@trusted-logic.fr>