SGX: UDD: Support all FBs reported by the FB driver.
The number of supported FB devices is currently hardcoded to 1.
This patch enables as many FB devices as is reported by the FB
driver.

Also, in case there is more than one FB device, they are
registered to the SGX driver in reverse order, as the SGX driver
will sequence them in reverse order. This makes sure, the FB device
number corresponds to the id assigned in services.

Change-Id: Ib4add1b3cfe292e7cea6f58e6204c6ee8256588e
Signed-off-by: Cemil Azizoglu <c-azizoglu@ti.com>