OMAP4: map OK key to ENTER.
Keys with a scancode bigger than 0x100,
are not generating a event because,
the keybit in the function is_event_supported
on drivers/input/input.c is not set.

Since android anyway maps the OK key to
ENTER,  workaround this problem by generating the
ENTER event until keys with scancode greater
than 256 are recognized.

Change-Id: Ic049ec05abe0c060eb1349b6e10c6251594ab218
Signed-off-by: Axel Haslam <axelhaslam@ti.com>