OMAP4 Board File: Change the LCD backlight interface to handle a value of 1
For the lcd backlight if a value of 1 was sent into the implementation the
algorithim indicates to divide by 2.  1/2 is 0.5 which translate to 0 and
there for makes the display come on at full intensity.

Since the display brightnesses value of 1 is pratically off then make
the display backlight off when 1 is sent.

Change-Id: Id089e21e706318c8807b602e83aadb5694757820
Signed-off-by: Dan Murphy <dmurphy@ti.com>