Keypad LEDs: Change the keypad backlight interface to handle a value of 1
For the keypad backlight if a value of 1 was sent into the implementation the
algorithim indicates to divide by 2.  1/2 is 0.5 which translate to 0 and
there for makes the display come on at full intensity.

Since the keypad brightnesses value of 1 is pratically off then make
the keypad backlight off when 1 is sent.

Change-Id: Iea48aace25c2546da33eb70a24568d0a2e6fa4b6
Signed-off-by: Dan Murphy <dmurphy@ti.com>