dspbridge: counter for reg/unreg object calls
The database configuration functions register only once
an object of the same type and UUID, this leads to problems
when there are multiple instances of the same node trying to
use the same object when deleting/creating nodes.

A counter has been created to maintain such references to
users of the registered object type, which should be enough
to keep consistency since we don't care to maintain who is
using what because it would add unnecessary complexity to
what should be a simple mechanism.

Change-Id: I0c20e767e6db31f520ed7a8b79d3ed2cbfe4db12
Signed-off-by: Omar Ramirez Luna <omar.ramirez@ti.com>