USB: MUSB: Do not enable TX and RX DMA at the same time
This is a workaround for OMAP3630 errata 1.84:
USB OTG DMA may stall under specific configuration.

The USB OTG module is AHB native, while the interconnect
is OCP native. This issue impacts the AHB/OCP wrapper
around the USB OTG module. It can only occur if at least
one Read DMA channel and one Write DMA channel are active
concurrently. This issue does not occur with only one DMA
channel active at a time.

With this workaround, in the best case, we effectively use
DMA in both directions, and in the worst case, we use PIO
mode in one direction.

Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>