OMAP3HQ: omap34xxcam: Use try_lock into omap34xxcam_poll function
According to documentation for linux device driver callback poll()
should never be blocked. Only need to attach its wait-queue and returns
mask with current available events. There is a case in which the
mutex vb_lock are locks for a long time. For example ioctl VIDIOC_DQBUF
can lock it for whole frame.

Signed-off-by: Dinko Mironov <dmironov@mm-sol.com>