TWL4030: BCI: Fix to reflect battery voltage change in sysfs
The battery voltage in sysfs of twl4030_bci_battery driver is not being
updated correctly. In case of no charger present the previous the voltage is
being retained in the registers of BCI module in twl4030. To fix this issue,
the values are being read from MADC module registers directly when no charger
in present.

Signed-off-by: Nagayoshi Inoue <inoue.nagayoshi@kk.jp.panasonic.com>
Signed-off-by: Jeetesh Burman <jeetesh.burman@ti.com>
Signed-off-by: Ponmuthu Subramaniam <ponmuthu.subramaniam@ti.com>