OMAP: MUX: Dont corrupt mux name
If we operate on the pointer passed,
we may corrupt the data section. Using
the same pointer to the same data section
would have the '.' replaced by a '\0'
which would mean that setting the mux mode
again would not work.

Instead on operating on the pointer recieved
as a parameter, we do a local copy that
can be discarded on exit.

Change-Id: 	I2057f1075a02d39527f07ba52ee8f1289fd5ae9a
Signed-off-by: Axel Haslam <axelhaslam@ti.com>
Reviewed-by: Praneeth Bajjuri <praneeth@ti.com>