OMAP: mailbox: write the mbox message to the hw when possible.
Currently when a new message is sent to the DSP it has to be
written first to a kfifo to schedule later a tarklet which will
do the actual transfer to the mailbox's hw fifo, also when the
tasklet tries to do the transfer it first polls for 1 ms in the
case that the hw fifo is full.
This patch allows the writing direclty to the mailbox's hw fifo
whenever possible avoiding the schedule of the tasklet. Also
when needed to put the message in the kfifo, the taskelt won't
spend the 1 ms trying to send the message, it will instead enable
the mailbox's full interrupt inmediately.

Change-Id: I9853ef020b9471468617a67278a8a78712f37db8
Signed-off-by: Rene Sapiens <rene.sapiens@ti.com>