OMAP3630: isp: Add function to control CCDC VP clock
1. Remove pixelclk filed from platform board configuration
because this is dinamic paramete depending on the current
sensor mode. And adding field in the main isp structure.

2. Add function isp_set_ccdc_vp_clock() to control and set
the new parameter, with appropriate controls for maximum and
minimum value. To ensure compatibility with old code value 0
means maximum VP clock.

3. Fix minimum clamp value in function isp_adjust_bandwidth of
divider. Correct minimum value is 2. Because a function call
ispccdc_config_vp_freq () from 'div' value is subtracted 2.

Signed-off-by: Dinko Mironov <dmironov@mm-sol.com>