OMAP3:PM:SDRC stabalization delay optimization
When 1.2G is enabled, the stablization delay computed for that
rate is causing GFX underflow. This implies, the DDR stall is more
than rate-at-which FIFO is getting drained.
As per experiements done on .29/.32 3430/3630 the delay of 11-12uSec
s/w delay is good enough for DDR stabalization.
Keeping all the data into consideration we can further optimize the
delay. This patch reduces the MPU delay loop by 6. This has been
validated on 3630 + froyo.

Change-Id: I6f51033bcac81170c517a3d924c77e9186729f7c
Signed-off-by: Girish S G <girishsg@ti.com>
Signed-off-by: Praneeth Bajjuri <praneeth@ti.com>