USB: MUSB: BUG-FIX: MUSB to work correctly after Suspen-Resume/idle
This patch fixes a long standing bug where idle path CORE OFF followed
by a suspend-resume cycle would lead to saving the wrong MUSB register
values and hence MUSB would not work properly after a idle/suspend-
resume cycles.

With this patch we always save the working configuration of NOSTDBY
and NOIDLE in each of the musb_save_context routines so that on a
restore we always restore the correct MUSB configuration.

With this patch we save a valid working context after MUSB init
and also remove the musb_save_context call from the suspend path
because if a idle path CORE OFF is hit before suspend, it will
result in a save-restore of wrong values.

Signed-off-by: Girish S G <girishsg@ti.com>
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>