USB: MUSB: Hold/Release a wakelock for MUSB on attach/detach
This patch holds a wakelock whenever a MUSB cable is attached
and will release the wakelock on a detach, thereby not allowing
suspend to happen when MUSB cable is connected. If a suspend
happens when a MUSB cable is connected it results in
un-predicatable behavior like board hang. This patch resolves
the above mentioned issue

Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>