dspbridge: change mmufault tasklet to a workqueue
Fix the warning appearing when printing the DSP's
stack. The warning is due to the fact that when Bridge
prints the mmu information it also enables the GPT7 clock,
the running in the context of the tasklet when calling
omap_dm_timer_enable(), makes the "__mutex_lock_slowpath"
to appear due to the usage of the hwmod's mutex.

Change-Id: Ie87f896654d3cfeb31cb86cb1d003b8de52f1c90
Signed-off-by: Fernando Guzman Lugo <x0095840@ti.com>
Signed-off-by: Rene Sapiens <rene.sapiens@ti.com>