dspbridge: unload dlls on node create failure
If a failure occurs in node create path, dsp dlls will remain
loaded forever (from bridge perspective), however if the
baseimage is reloaded it will clean the module information of
the previously loaded dlls, next use case will avoid reloading
the common dlls and also overwrite the current code in the
shared memory causing unexpected behavior.

This is solved by unloading the code right after a failure (and
if loaded in the first place).

Signed-off-by: Omar Ramirez Luna <omar.ramirez@ti.com>