OMAP3: Camera: Add missing call to stop ISP in 720p path
After board boot up, go to Gallery and play a 720p clip.
Then go to OMAP3 camera app and start preview. A blank
screen will be observed. This patch fixes this issue.

During 720p playback, isp_start() is called which updates
isp->running to ISP_RUNNING.

Now, when we start preview, during configuration for AF
and H3A, ispstat_bufs_alloc() is called which expects
isp->running to be in the ISP_STOPPED state.

In the 720p path, ispdss_put_resource() is called to release
all resources and was missing a call to isp_stop(), which
would update isp->running to ISP_STOPPED.

Change-Id: I742c9125bd1914170be6c678b63879f46d3b534d
Signed-off-by: Christina Warren <cawarren@ti.com>