I2C: Fix crash during I2C timeout
During stress test, on one of our customer platform
we saw a I2C timeout which was leading to crash. I2C
write was initiated by twl4030_madc_write() function
in twl4030-madc.c file.

In case of twl_i2c_write_u8() write failure, dev_err
is called but the structure madc->dev was found to be
null which was leading to crash. On debugging it was
found that madc->dev was not filled in the
twl4030_madc_probe().

This patch fixes the issue.

Change-Id: I0110cee0c5d73d9f0c72907616d8cb1742de2ac6
Signed-off-by: Narasimha Gururaj <x0083234@ti.com>
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>