HSMMC:Allow time for card detect debouncing.
The 8ms and 30ms debounce time in controller and the gpio detect
signal respectively is not enough to handle a use case where a
user removes the card slowly from the cage.A failure is noticed when
the card is removed slowly due to the fact that card is still on the
bus resulting in status cmds to sucess.

This patch introduces resonable delays of 0.5 and 1.5 secs to handle
this in card plug/unplug use cases.

Signed-off-by: Madhusudhan Chikkature <madhu.cr@ti.com>
Reviewed-by: Praneeth Bajjuri <praneeth@ti.com>