USB: MUSB: Save initial musb-context at right place during bootup
This patch solves a BUG wherein during bootup the musb context
was saved at a wrong place which was resulting in the loss of
SOFT CONNECT bit for the case when CORE hits OFF before a usb
cable is attached after bootup.

Although, this issue is hidden on the Zoom boards most of the
time because of a spurious attach/detach interrupt, which saves
a working context at the end of the bootup. But this spurious
interrupts do not happend 2 in 10 times thereby exposing the
mentioned issue. This patch fixes this issue.

Also, in this patch move the restore-context call within the
phy_resume function, above the twl-asleep check, so that the
context is re-stored first time when the CORE hits OFF before
a usb cable is attached eventhough asleep=0 after musb_init
during bootup.

Fix for OMAPS00229886

Change-Id: I07596d8098ea7eeaa5a69b4329da7f74070069c8
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>
Reviewed-by: Dan Murphy <dmurphy@ti.com>
Acked-by: Praneeth Bajjuri <praneeth@ti.com>